implement error models to report commands

+++ src/GasStation.Application/Commands/Report/Delete/DeleteReportCommandHandler.cs
@@ -1,10 +1,12 @@
 using GasStation.Application.Common.Interfaces.Persistence;
 using MediatR;
+using ErrorOr;
+using GasStation.Domain.Errors;
 using Microsoft.EntityFrameworkCore;
 
 namespace GasStation.Application.Commands.Report.Delete;
 
-public class DeleteReportCommandHandler : IRequestHandler<DeleteReportRequest, DeleteReportResponse>
+public class DeleteReportCommandHandler : IRequestHandler<DeleteReportRequest, ErrorOr<DeleteReportResponse>>
 {
     private readonly IApplicationDbContext _dbContext;
 
@@ -13,16 +15,19 @@ public class DeleteReportCommandHandler : IRequestHandler<DeleteReportRequest, D
         _dbContext = dbContext;
     }
     
-    public async Task<DeleteReportResponse> Handle(DeleteReportRequest request, CancellationToken cancellationToken)
+    public async Task<ErrorOr<DeleteReportResponse>> Handle(DeleteReportRequest request, CancellationToken cancellationToken)
     {
         var report = await _dbContext.Reports
-                         .FirstOrDefaultAsync(r => r.Title == request.Title, cancellationToken)
-                         ?? throw new Exception("The report with specified title doesn't exist!");
+                         .FirstOrDefaultAsync(r => r.Title == request.Title, cancellationToken);
+        if (report is null)
+        {
+            return Errors.Report.TitleNotFound;
+        }
         
         _dbContext.Reports.Remove(report);
         var result = await _dbContext.SaveChangesAsync(cancellationToken);
         
         return result > 0 ? new DeleteReportResponse() {IsDeleted = true} 
-            : throw new Exception("Something went wrong!");
+            : Errors.Database.Unexpected;
     }
 }
